+++ rutui.gemspec
@@ -7,7 +7,7 @@ Gem::Specification.new do |s|
   s.license     = "MIT"
   s.authors     = ["Roman Pramberger"]
   s.email       = 'roman@pramberger.ch'
-  s.files       = [ "lib/rutui.rb", "lib/rutui/util.rb", "lib/rutui/pixel.rb", 
+  s.files       = [ "lib/rutui.rb", "lib/rutui/ansi.rb", "lib/rutui/pixel.rb",
   					"lib/rutui/theme.rb", "lib/rutui/objects.rb", "lib/rutui/screen.rb", 
   					"lib/rutui/screenmanager.rb", "lib/rutui/figlet.rb", "lib/rutui/axx.rb", 
   					"lib/rutui/sprites.rb", "lib/rutui/table.rb", "lib/rutui/textfield.rb"]
